requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
email-validator
python-multipart
